backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
uvicorn==0.23.2
torch==2.0.1
transformers==4.30.2
tokenizers==0.13.3
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
readability-lxml==0.8.4.1
lxml==4.9.3
keybert==0.7.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
flask-limiter==3.5.0
pytest==7.4.2
pytest-flask==1.2.0
datasets==2.13.0
matplotlib==3.7.2
huggingface-hub==0.16.4
pyarrow==12.0.1
accelerate==0.20.3
peft==0.4.0 